Allow to keep test tempdirs by setting RNP_KEEP_TEMP env var

+++ src/tests/ffi.cpp
@@ -529,7 +529,9 @@ TEST_F(rnp_tests, test_ffi_save_keys)
     rnp_ffi_destroy(ffi);
 
     // final cleanup
-    delete_recursively(temp_dir);
+    if (!getenv("RNP_KEEP_TEMP")) {
+        delete_recursively(temp_dir);
+    }
     free(temp_dir);
 }
 
@@ -617,7 +619,9 @@ TEST_F(rnp_tests, test_ffi_load_save_keys_to_utf8_path)
     rnp_ffi_destroy(ffi);
 
     // final cleanup
-    delete_recursively(temp_dir);
+    if (!getenv("RNP_KEEP_TEMP")) {
+        delete_recursively(temp_dir);
+    }
     free(temp_dir);
 }
