+++ src/tests/rnp_tests.cpp
@@ -74,7 +74,9 @@ rnp_tests::rnp_tests() : m_dir(make_temp_dir())
 
 rnp_tests::~rnp_tests()
 {
-    delete_recursively(m_dir);
+    if (!getenv("RNP_KEEP_TEMP")) {
+        delete_recursively(m_dir);
+    }
     free(m_dir);
 }
